Move syntax to separate file, fix `sbcomp`.
Substitution *acts on the right* in the sense that `A (σ ∘ τ) ≡ (A σ) τ)`.
With the current definition of `sbcomp` we get instead `A (σ ∘ τ) ≡ (A τ) σ)`
which corresponds to a *left* action. So, either we change the notation
for the action of substitutio or we change `sbcomp`. The choice is clear,
we need to change `sbcomp` and be compatible with the rest of mathematics
(and who cares about crazy computer scientists that write composition
the wrong way).