Improve magic to 'push' substitutions inside variables
Up till now it only took care of `sbzero` and `sbshift` for `var 0` but
this adds support for `sbweak` for any `var k` and `sbzero` for `var (S
k)`.
It’s still missing the rule for `sbshift` and there might be problems
as `var 1` might not match `var (S ?k)` without some work.